Create Library
This file is the library, it has functions and you can add to it. It also has functions to read and write from the magnetometer. I was following a tutorial for i2c at the time so that may be why its familiar but you can just use the normal functions if you want